backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "crypto-investor"
version = "0.1.0"
description = "Crypto investment platform backend"
requires-python = ">=3.10"
dependencies = [
    "argon2-cffi>=23,<24",
    "ccxt>=4,<5",
    "channels>=4,<5",
    "cryptography>=44,<45",
    "daphne>=4,<5",
    "django>=5.1,<6",
    "django-cors-headers>=4,<5",
    "djangorestframework>=3.15,<4",
    "httpx>=0.27,<1",
    "pydantic>=2,<3",
    "python-dotenv>=1,<2",
]

[project.optional-dependencies]
analysis = [
    "pandas>=2,<3",
    "numpy>=1.26,<2",
    "scipy>=1.11,<2",
]
dev = [
    "pytest>=8,<9",
    "pytest-django>=4,<5",
    "pytest-asyncio>=0.24,<1",
    "httpx>=0.27,<1",
    "ruff>=0.8,<1",
    "mypy>=1.13,<2",
    "pip-audit>=2,<3",
]

[tool.hatch.build.targets.wheel]
packages = ["config"]

[tool.ruff]
target-version = "py310"
line-length = 100
exclude = ["*/migrations/*"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "A", "SIM"]

[tool.ruff.lint.per-file-ignores]
"market/services/regime.py" = ["E402"]
"tests/*.py" = ["E402"]

[tool.mypy]
python_version = "3.10"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
DJANGO_SETTINGS_MODULE = "config.settings"
pythonpath = ["."]
